requirements.txt
# Car Brand Detection - Windows Requirements
# Install with: pip install -r requirements.txt

# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Image processing
opencv-python>=4.8.0
Pillow>=9.5.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter (optional)
jupyter>=1.0.0
notebook>=6.5.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
pathlib2>=2.3.7

# Windows specific
pywin32>=306; sys_platform == "win32"